requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
pipreqs==0.4.13
django-cors-headers==4.0.0